Working Twitter bot with some issues that needs to be fixed

+++ database/migrations/2017_03_30_154253_create_trackings_table.php
@@ -15,7 +15,7 @@ class CreateTrackingsTable extends Migration
     {
         Schema::create('trackings', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('twitter_id')->unsigned();
+            $table->bigInteger('twitter_id')->unsigned();
             $table->timestamps();
         });
     }
